Agregando Sweet Alert al proyecto

+++ northwind-app/src/app/modules/categories/components/categories-table/categories-table.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
+
 import { Category } from '../../models/Category';
 import { CategoriesService } from '../../services/categories.service';
 
@@ -22,9 +24,29 @@ export class CategoriesTableComponent implements OnInit {
   }
 
   onDelete(id: number) {
-    this.categoriesService.delete(id).subscribe(res => {
-      this.categoriesList = this.categoriesList.filter(c => c.id != id);
-    });
+    Swal.fire({
+      title: '¿Estás seguro que deseas eliminar la categoría?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#22c55e',
+      cancelButtonColor: '#dc2626',
+      confirmButtonText: 'Si, eliminar!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.categoriesService.delete(id).subscribe(res => {
+          Swal.fire({
+            title: '¡Eliminado!',
+            text: 'La categoría se ha eliminado correctamente',
+            icon: 'success',
+            confirmButtonColor: '#22c55e'
+          });
+
+          this.categoriesList = this.categoriesList.filter(c => c.id != id);
+        });
+      }
+    })
+    
   }
 
 }
